Extract currency select hook and show bulk media count.
Move currency dropdown state into useCurrencySelect, and label the bulk inspector with the selected media count.
